+++ cypress/integration/demo.spec.ts
@@ -10,6 +10,33 @@ describe('Rule editor demo', () => {
       cy.get('lhc-rule-editor').should('not.exist');
     });
 
+    it('should disable the "Open Rule Editor" button if questionnaire/Root level/Question have not been selected', () => {
+      cy.get('select#questionnaire-select').select('Upload your own questionnaire');
+
+      // The 'Open Rule Editor' should be disabled
+      cy.get('#openRuleEditor').should('have.class', 'disabled');
+
+      // Select file to upload
+      cy.get('#file-upload').attachFile('bmi.json');
+
+      // By default, the 'Root level' checkbox is checked
+      cy.get('#root-level').should('exist').should('be.checked')
+      // The 'Open Rule Editor' should be enabled in this case
+      cy.get('#openRuleEditor').should('not.have.class', 'disabled');
+
+      // Uncheck the 'Root level' checkbox
+      cy.get('#root-level').uncheck();
+      // The 'Open Rule Editor' should be disabled
+      cy.get('#openRuleEditor').should('have.class', 'disabled');
+
+      // Select Question - BMI
+      cy.get('#question').type('bmi');
+      cy.get('span#completionOptions > ul > li').contains('39156-5').click();
+      // The 'Open Rule Editor' should be enabled again.
+      cy.get('#openRuleEditor').should('not.have.class', 'disabled');
+
+    });
+
     it('should successfully upload a JSON file', () => {
       cy.get('select#questionnaire-select').select('Upload your own questionnaire');