+++ projects/ng-rule-editor/src/lib/rule-editor.component.spec.ts
@@ -8,6 +8,7 @@ import { CaseStatementsComponent } from './case-statements/case-statements.compo
 import { MatSnackBarModule } from '@angular/material/snack-bar';
 import { SyntaxPreviewComponent } from './syntax-preview/syntax-preview.component';
 import { MatTooltipModule } from '@angular/material/tooltip';
+import { ValidationResult } from './variable';
 
 describe('RuleEditorComponent', () => {
   let component: RuleEditorComponent;
@@ -35,5 +36,67 @@ describe('RuleEditorComponent', () => {
 
   it('should create', () => {
     expect(component).toBeTruthy();
+  });
+
+  it('should return empty aria message if there is no error in the validation result', () => {
+    const validationResult: ValidationResult = {
+      hasError: false,
+      errorInItemVariables: false,
+      errorInOutputExpression: false,
+      errorInOutputCaseStatement: false
+    }
+
+    expect(component.composeAriaValidationErrorMessage(validationResult)).toEqual("");
+  });
+
+  it('should return an aria message if there is an error in the Item Variable section', () => {
+    const validationResult: ValidationResult = {
+      hasError: true,
+      errorInItemVariables: true,
+      errorInOutputExpression: false,
+      errorInOutputCaseStatement: false
+    }
+
+    expect(component.composeAriaValidationErrorMessage(validationResult))
+           .toEqual("The 'save' button is disabled due to one or more errors in the Item Variable section.");
+  });
+
+  it('should return an aria message if there is an error with the expression in the Output Expression section', () => {
+    const validationResult: ValidationResult = {
+      hasError: true,
+      errorInItemVariables: false,
+      errorInOutputExpression: true,
+      errorInOutputCaseStatement: false
+    }
+
+    expect(component.composeAriaValidationErrorMessage(validationResult))
+           .toEqual("The 'save' button is disabled due to one or more errors with the expression in the" +
+           " Output Expression section.");
+  });
+
+  it('should return an aria message if there is one or more errors with the case statement in the Output Expression section', () => {
+    const validationResult: ValidationResult = {
+      hasError: true,
+      errorInItemVariables: false,
+      errorInOutputExpression: false,
+      errorInOutputCaseStatement: true
+    }
+
+    expect(component.composeAriaValidationErrorMessage(validationResult))
+           .toEqual("The 'save' button is disabled due to one or more errors with the case statement in the" +
+           " Output Expression section.");
+  });
+
+  it('should return an aria message if there are errors in Item Variables and the Output Expression sections', () => {
+    const validationResult: ValidationResult = {
+      hasError: true,
+      errorInItemVariables: true,
+      errorInOutputExpression: true,
+      errorInOutputCaseStatement: false
+    }
+
+    expect(component.composeAriaValidationErrorMessage(validationResult))
+           .toEqual("The 'save' button is disabled due to errors in the Item Variable section, and" +
+           " with the expression in the Output Expression section.");
   });
 });