+++ CHANGELOG.md
@@ -2,13 +2,18 @@
 
 This project follows [Semantic Versioning](http://semver.org/).
 
-## [3.1.19] 2024-04-04
+## [3.1.20] 2024-05-03
 ### Changed
+- Enable the 'Open Rule Editor' button only if the questionnaire is selected and either the 'Root level' checkbox or Question drop-down is selected.
 - Allowed FHIR Query with parameters to be parsed in any order
+
+## [3.1.19] 2024-04-01
+### Added
+- Added variable name validation
 
 ## [3.1.18] 2024-02-27
 ### Changed
-- Added variable name validation
+- Added a screen reader error message for variables used in the output expression
 
 ## [3.1.17] 2024-01-26
 ### Added